requirements.txt
opencv-python
numpy
scenedetect
pygame
pillow
